+++ src/main/kotlin/kr/mashup/bangwidae/asked/model/question/Question.kt
@@ -27,6 +27,8 @@ data class Question(
     val representativeAddress: String? = null,
     val region: Region? = null,
 
+    // TODO anonymous 기존 데이터 마이그레이션 후 not-null 로 수정
+    val anonymous: Boolean? = false,
     val deleted: Boolean = false,
     @Version var version: Int? = null,
     @CreatedDate var createdAt: LocalDateTime? = null,
